+++ src/ha/load-components.ts
@@ -8,6 +8,7 @@ const HA_COMPONENTS = [
   "ha-control-circular-slider",
   "ha-control-button",
   "ha-icon-button",
+  "ha-color-picker",
   "ha-md-list",
   "ha-md-list-item",
   "ha-input-chip"